displayLibrary and removeBook added
